# api-gateway/api/docs/swagger.yaml
definitions:
  models.CreateCommentRequest:
    properties:
      description:
        type: string
      liked:
        type: boolean
      post_id:
        type: string
      user_id:
        type: string
    type: object
  models.CreateCommentResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      liked:
        type: boolean
      post_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.CreatePostRequest:
    properties:
      comment:
        $ref: '#/definitions/models.CreateCommentRequest'
      id:
        type: string
      photo:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  models.CreatePostResponse:
    properties:
      comment:
        $ref: '#/definitions/models.CreateCommentResponse'
      created_at:
        type: string
      id:
        type: string
      photo:
        type: string
      title:
        type: string
      user_id:
        type: string
    type: object
  models.CreateRegister:
    properties:
      code:
        type: string
      email:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  models.CreateUserRequest:
    properties:
      code:
        type: string
      email:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      password:
        type: string
      post:
        $ref: '#/definitions/models.CreatePostRequest'
      refresh_token:
        type: string
      username:
        type: string
    type: object
  models.CreateUserResponse:
    properties:
      access_token:
        type: string
      code:
        type: string
      email:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      password:
        type: string
      post:
        $ref: '#/definitions/models.CreatePostResponse'
      refresh_token:
        type: string
      username:
        type: string
    type: object
  models.Empty:
    type: object
  models.Error:
    properties:
      code:
        type: integer
      description:
        type: string
      error: {}
    type: object
  models.GetAllUserRequest:
    properties:
      limit:
        type: integer
      page:
        type: integer
    type: object
  models.GetAllUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
      refresh_token:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.GetAllUsersResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/models.GetAllUserResponse'
        type: array
    type: object
...
  models.GetUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
      refresh_token:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      last_name:
        type: string
      name:
        type: string
      username:
        type: string
    type: object
  models.UpdateUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      last_name:
        type: string
      name:
        type: string
      refresh_token:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  models.VerifyResponse:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      password:
        type: string
      refresh_token:
        type: string
      username:
        type: string
    type: object
host: localhost:7000
info:
  contact: {}
  title: Go-BootCamp N7
  version: "1.0"
paths:
...
  /v1/register:
    post:
      consumes:
      - application/json
      description: this registers customer
      parameters:
      - description: Register customer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateRegister'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: register customer
      tags:
      - register
  /v1/user:
    delete:
      consumes:
      - application/json
      description: This API Can Delete User
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Empty'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete User
      tags:
      - user
    get:
      consumes:
      - application/json
      description: This API Can Get User Info
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GetUserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get User
      tags:
      - user
    post:
      consumes:
      - application/json
      description: This API Can Create User
      parameters:
      - description: Get Users By LImit Page
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateUserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: CrateUser
      tags:
      - user
    put:
      consumes:
      - application/json
      description: This API Can Update All Informations
      parameters:
      - description: UpdateUser
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UpdateUserResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - user
  /v1/users:
    get:
      consumes:
      - application/json
      description: This API Can Get Users By Limit
      parameters:
      - description: Get Users By LImit Page
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.GetAllUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.GetAllUsersResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: GetUser
      tags:
      - user
  /v1/verify/{email}/{code}:
    get:
      consumes:
      - application/json
      description: this registers customer
      parameters:
      - description: email
        in: path
        name: email
        required: true
        type: string
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.VerifyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Error'
      summary: register customer
      tags:
      - register
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"